// src/modules/food/dto/search-recipes-request.dto.ts
export interface SearchRecipesRequest {
  results: Result[];
  offset: number;
  number: number;
  totalResults: number;
}

export interface Result {
  vegetarian: boolean;
  vegan: boolean;
  glutenFree: boolean;
  dairyFree: boolean;
  veryHealthy: boolean;
  cheap: boolean;
  veryPopular: boolean;
  sustainable: boolean;
  lowFodmap: boolean;
  weightWatcherSmartPoints: number;
  gaps: Gaps;
  preparationMinutes: number;
  cookingMinutes: number;
  aggregateLikes: number;
  healthScore: number;
  creditsText: Author;
  license: License;
  sourceName: SourceName;
  pricePerServing: number;
  id: number;
  title: string;
  readyInMinutes: number;
  servings: number;
  sourceUrl: string;
  image: string;
  imageType: ImageType;
  nutrition: Nutrition;
  summary: string;
  cuisines: string[];
  dishTypes: DishType[];
  diets: string[];
  occasions: string[];
  analyzedInstructions: AnalyzedInstruction[];
  spoonacularSourceUrl: string;
  author?: Author;
}

export interface AnalyzedInstruction {
  name: string;
  steps: Step[];
}
// ...
export enum Author {
  Coffeebean = 'coffeebean',
  FoodistaCOMTheCookingEncyclopediaEveryoneCanEdit = 'Foodista.com – The Cooking Encyclopedia Everyone Can Edit',
  Maplewoodroad = 'maplewoodroad'
}

export enum DishType {
  Dinner = 'dinner',
  Lunch = 'lunch',
  MainCourse = 'main course',
  MainDish = 'main dish',
  SideDish = 'side dish',
  Soup = 'soup'
}

export enum Gaps {
  No = 'no'
}

export enum ImageType {
  Jpg = 'jpg'
}

export enum License {
  CcBy30 = 'CC BY 3.0',
  SpoonacularSTerms = "spoonacular's terms"
}

export interface Nutrition {
  nutrients: Flavonoid[];
  properties: Flavonoid[];
  flavonoids: Flavonoid[];
  ingredients: Ingredient[];
  caloricBreakdown: CaloricBreakdown;
  weightPerServing: WeightPerServing;
}
// ...
export enum SourceName {
  Foodista = 'Foodista',
  MaplewoodRoad = 'Maplewood Road',
  Spoonacular = 'spoonacular'
}